Update type exports
