+++ src/SmartList/SmartList.tsx
@@ -1,8 +1,28 @@
 import React from 'react'
 
 import { Header } from './components/Header'
-import { IProps, IRowElement, IHeaderItem, IColumnElement } from './types'
 import { ListItem } from './components/ListItem'
+
+export interface IProps {
+  data: IRowElement[]
+  cols: IHeaderItem[]
+}
+
+export interface IHeaderItem {
+  name: string
+  className?: string
+  handleSort?: (column: string) => void
+}
+
+export interface IColumnElement {
+  name: string
+  element: JSX.Element
+}
+
+export interface IRowElement {
+  id: number | string
+  columns: IColumnElement[]
+}
 
 const SmartList: React.FC<IProps> = ({ data, cols }) => {
